[closes #14] Map Entity 설계
