+++ src/main/java/com/muze/domain/map/command/domain/aggregate/vo/MemberVO.java
@@ -1,0 +1,22 @@
+package com.muze.domain.map.command.domain.aggregate.vo;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Getter
+@Embeddable
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MemberVO {
+    @Column(nullable = false, name = "member_id")
+    private Long id;
+
+    @Builder
+    public MemberVO(long memeberId){
+        this.id=memeberId;
+    }
+}
