--- src/main/java/com/muze/domain/security/command/application/controller/AuthController.java
@@ -1,6 +1,0 @@
-package com.muze.domain.security.command.application.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-public class AuthController {
-}
