--- src/main/java/com/muze/domain/security/command/domain/exception/BadRequestException.java
@@ -1,11 +1,0 @@
-package com.muze.domain.security.command.domain.exception;
-
-public class BadRequestException extends RuntimeException{
-    public BadRequestException(String message) {
-        super(message);
-    }
-
-    public BadRequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
