--- src/main/java/com/muze/domain/security/command/domain/exception/OAuth2AuthenticationProcessingException.java
@@ -1,13 +1,0 @@
-package com.muze.domain.security.command.domain.exception;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class OAuth2AuthenticationProcessingException extends AuthenticationException {
-    public OAuth2AuthenticationProcessingException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-
-    public OAuth2AuthenticationProcessingException(String msg) {
-        super(msg);
-    }
-}
