--- src/main/java/com/muze/domain/security/command/domain/provider/OAuth2UserInfo.java
@@ -1,25 +1,0 @@
-package com.muze.domain.security.command.domain.provider;
-
-import java.util.Map;
-
-public abstract class OAuth2UserInfo {
-    protected Map<String, Object> attributes;
-
-    public OAuth2UserInfo(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    public abstract String getProvider();
-
-    public abstract String getId();
-
-    public abstract String getName();
-
-    public abstract String getEmail();
-
-    public abstract String getImageUrl();
-}
