--- src/main/java/com/muze/domain/security/command/domain/service/RequestMemberService.java
@@ -1,7 +1,0 @@
-package com.muze.domain.security.command.domain.service;
-
-import com.muze.domain.member.query.application.dto.FindMemberDTO;
-
-public interface RequestMemberService {
-    FindMemberDTO findMemberById(Long memberId);
-}
