+++ src/main/java/com/muze/global/configration/SecurityConfiguration.java
@@ -1,8 +1,9 @@
 package com.muze.global.configration;
 
 import com.muze.domain.member.command.domain.aggregate.entity.enumtype.Role;
-import com.muze.domain.security.command.application.service.CustomOAuth2UserService;
-import com.muze.domain.security.command.domain.repository.HttpCookieOAuth2AuthorizationRequestRepository;
+import com.muze.global.security.command.application.service.CustomOAuth2UserService;
+import com.muze.global.security.command.domain.repository.HttpCookieOAuth2AuthorizationRequestRepository;
+import com.muze.global.handler.CustomAccessDeniedHandler;
 import com.muze.global.handler.CustomOAuth2FailureHandler;
 import com.muze.global.handler.CustomOAuth2SucessHandler;
 import lombok.RequiredArgsConstructor;
@@ -22,13 +23,18 @@ public class SecurityConfiguration {
     private final CustomOAuth2UserService customOAuth2UserService;
     private final CustomOAuth2SucessHandler customOAuth2SucessHandler;
     private final CustomOAuth2FailureHandler customOAuth2FailureHandler;
-//    @Autowired
-//    @Qualifier("RestAuthenticationEntryPoint")
-//    private AuthenticationEntryPoint authEntryPoint;
+
+    private final CustomAccessDeniedHandler customAccessDeniedHandler;
+
+    @Autowired
+    @Qualifier("RestAuthenticationEntryPoint")
+    private AuthenticationEntryPoint authEntryPoint;
 
     public HttpCookieOAuth2AuthorizationRequestRepository httpCookieOAuth2AuthorizationRequestRepository(){
         return new HttpCookieOAuth2AuthorizationRequestRepository();
     }
+
+
 
     @Bean
     public SecurityFilterChain filterChain (HttpSecurity http) throws Exception{
@@ -43,7 +49,7 @@ public class SecurityConfiguration {
                 .formLogin()
                 .disable()
                 .exceptionHandling()
-//                .authenticationEntryPoint(authEntryPoint)
+                .authenticationEntryPoint(authEntryPoint)
 
                 .and()
 
@@ -69,7 +75,9 @@ public class SecurityConfiguration {
                 .successHandler(customOAuth2SucessHandler)
                 .failureHandler(customOAuth2FailureHandler);
 
-
+        http
+                .exceptionHandling()
+                .accessDeniedHandler(customAccessDeniedHandler);
         return http.build();
     }
 }
