+++ src/main/java/com/muze/global/security/command/application/controller/AuthController.java
@@ -1,0 +1,6 @@
+package com.muze.global.security.command.application.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+public class AuthController {
+}
