+++ src/main/java/com/muze/global/security/command/domain/exception/BadRequestException.java
@@ -1,0 +1,11 @@
+package com.muze.global.security.command.domain.exception;
+
+public class BadRequestException extends RuntimeException{
+    public BadRequestException(String message) {
+        super(message);
+    }
+
+    public BadRequestException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
