+++ src/main/java/com/muze/global/security/command/domain/exception/TokenNotFoundException.java
@@ -1,0 +1,13 @@
+package com.muze.global.security.command.domain.exception;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class TokenNotFoundException extends AuthenticationException {
+    public TokenNotFoundException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+    public TokenNotFoundException(String msg) {
+        super(msg);
+    }
+}
