+++ src/main/java/com/muze/global/security/command/domain/provider/Kakao.java
@@ -1,0 +1,50 @@
+package com.muze.global.security.command.domain.provider;
+
+import java.util.Map;
+
+public class Kakao extends OAuth2UserInfo{
+    public Kakao(Map<String, Object> attributes) {
+        super(attributes);
+    }
+
+    @Override
+    public String getId() {
+        return attributes.get("id").toString();
+    }
+
+    @Override
+    public String getName() {
+        Map<String, Object> properties = (Map<String, Object>) attributes.get("properties");
+
+        if (properties == null) {
+            return null;
+        }
+
+        return (String) properties.get("nickname");
+    }
+
+    @Override
+    public String getEmail() {
+        Map<String, Object> properties = (Map<String, Object>) attributes.get("kakao_account");
+        if (properties == null) {
+            return null;
+        }
+        return (String) properties.get("email");
+    }
+
+    @Override
+    public String getImageUrl() {
+        Map<String, Object> properties = (Map<String, Object>) attributes.get("properties");
+
+        if (properties == null) {
+            return null;
+        }
+
+        return (String) properties.get("thumbnail_image");
+    }
+
+    @Override
+    public String getProvider(){
+        return "KAKAO";
+    }
+}
