+++ src/main/java/com/muze/global/security/command/domain/provider/OAuth2UserInfoFactory.java
@@ -1,0 +1,13 @@
+package com.muze.global.security.command.domain.provider;
+
+import java.util.Map;
+
+public class OAuth2UserInfoFactory {
+    public static OAuth2UserInfo getOAuth2UserInfo(String registrationId, Map<String, Object> attributes) {
+        if (registrationId.equalsIgnoreCase("KAKAO")) {
+            return new Kakao(attributes);
+        } else {
+            throw new IllegalArgumentException("해당 OAuth2 제공자는 지원하지 않습니다");
+        }
+    }
+}
