+++ src/main/java/com/muze/global/security/command/domain/repository/TokenRepository.java
@@ -1,0 +1,11 @@
+package com.muze.global.security.command.domain.repository;
+
+import com.muze.global.security.command.domain.aggregate.entity.Token;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface TokenRepository extends JpaRepository<Token, Long> {
+    Optional<Token> findTokenByMember_Id(long memberId);
+    Optional<Token> findTokenByAccessToken(String accessToken);
+}
