+++ src/main/java/com/muze/global/security/command/domain/service/RequestMemberService.java
@@ -1,0 +1,7 @@
+package com.muze.global.security.command.domain.service;
+
+import com.muze.domain.member.query.application.dto.FindMemberDTO;
+
+public interface RequestMemberService {
+    FindMemberDTO findMemberById(Long memberId);
+}
